SW-26398 - Prepare commands and components for PHPStan update

+++ engine/Shopware/Components/MultiEdit/Resource/Product/Backup.php
@@ -520,7 +520,7 @@ class Backup
         $backup->setItems($items);
         $backup->setPath($path);
         $backup->setHash(sha1_file($path));
-        $backup->setSize(filesize($path));
+        $backup->setSize((int) filesize($path));
 
         $backup->setDate(new DateTime());
